requirements.txt
celery>=5.3.0
redis>=5.0.0
langchain>=0.1.0
langgraph>=0.0.10
langchain-openai>=0.0.5
python-dotenv>=1.0.0
